`check!`: report spec failures as literals when possible
Reporting the faulty specs as `'some-variable` (quoted symbol) tends to be less useful that reporting them as `::the-actual-keyword` (the keyword which is the actual spec being checked).

This optimization in reporting is only possible for keywords and sets, since functions, spec objects, etcc are more opaque (`'some-symbol` will be _more_ informative than a function representation in this case).

In case the optimization is applied, the quoted symbol will still be reported as an extra `:spec-quoted` key. Else, a `:raw-spec` will be associated, containing the spec object.